Struct google_appengine1_beta5::AppMethods

pub struct AppMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on app resources. It is not used directly, but through the Appengine hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_appengine1_beta5 as appengine1_beta5;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use appengine1_beta5::Appengine;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Appengine::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)`, `operations_get(...)`, `operations_list(...)`, `services_delete(...)`, `services_get(...)`, `services_list(...)`, `services_patch(...)`, `services_versions_create(...)`, `services_versions_delete(...)`, `services_versions_get(...)` and `services_versions_list(...)`
// to build up your call.
let rb = hub.apps();

Methods

impl<'a, C, A> AppMethods<'a, C, A>

fn services_list(&self, apps_id: &str) -> AppServiceListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all the services in the application.

Arguments

appsId - Part of name. Name of the resource requested. For example: "apps/myapp".

fn services_versions_create(&self, request: Version, apps_id: &str, services_id: &str) -> AppServiceVersionCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Deploys new code and resource files to a version.

Arguments

request - No description provided.

appsId - Part of name. Name of the resource to update. For example: "apps/myapp/services/default".

servicesId - Part of name. See documentation of appsId.

fn services_get(&self, apps_id: &str, services_id: &str) -> AppServiceGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the current configuration of the service.

Arguments

appsId - Part of name. Name of the resource requested. For example: "apps/myapp/services/default".

servicesId - Part of name. See documentation of appsId.

fn services_versions_list(&self, apps_id: &str, services_id: &str) -> AppServiceVersionListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the versions of a service.

Arguments

appsId - Part of name. Name of the resource requested. For example: "apps/myapp/services/default".

servicesId - Part of name. See documentation of appsId.

fn operations_get(&self, apps_id: &str, operations_id: &str) -> AppOperationGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

Arguments

appsId - Part of name. The name of the operation resource.

operationsId - Part of name. See documentation of appsId.

fn operations_list(&self, apps_id: &str) -> AppOperationListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding below allows API services to override the binding to use different resource name schemes, such as users/*/operations.

Arguments

appsId - Part of name. The name of the operation collection.

fn services_versions_delete(&self, apps_id: &str, services_id: &str, versions_id: &str) -> AppServiceVersionDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes an existing version.

Arguments

appsId - Part of name. Name of the resource requested. For example: "apps/myapp/services/default/versions/v1".

servicesId - Part of name. See documentation of appsId.

versionsId - Part of name. See documentation of appsId.

fn services_patch(&self, request: Service, apps_id: &str, services_id: &str) -> AppServicePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the configuration of the specified service.

Arguments

request - No description provided.

appsId - Part of name. Name of the resource to update. For example: "apps/myapp/services/default".

servicesId - Part of name. See documentation of appsId.

fn get(&self, apps_id: &str) -> AppGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets information about an application.

Arguments

appsId - Part of name. Name of the application to get. For example: "apps/myapp".

fn services_delete(&self, apps_id: &str, services_id: &str) -> AppServiceDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a service and all enclosed versions.

Arguments

appsId - Part of name. Name of the resource requested. For example: "apps/myapp/services/default".

servicesId - Part of name. See documentation of appsId.

fn services_versions_get(&self, apps_id: &str, services_id: &str, versions_id: &str) -> AppServiceVersionGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets application deployment information.

Arguments

appsId - Part of name. Name of the resource requested. For example: "apps/myapp/services/default/versions/v1".

servicesId - Part of name. See documentation of appsId.

versionsId - Part of name. See documentation of appsId.