Struct google_appengine1_beta5::Application

pub struct Application {
    pub default_hostname: Option<String>,
    pub name: Option<String>,
    pub code_bucket: Option<String>,
    pub default_bucket: Option<String>,
    pub dispatch_rules: Option<Vec<UrlDispatchRule>>,
    pub default_cookie_expiration: Option<String>,
    pub location: Option<String>,
    pub auth_domain: Option<String>,
    pub id: Option<String>,
}

An Application contains the top-level configuration of an App Engine application.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get apps (response)

Fields

default_hostname

The hostname used to reach the application, as resolved by App Engine. @OutputOnly

name

The full path to the application in the API. Example: "apps/myapp". @OutputOnly

code_bucket

A Google Cloud Storage bucket which can be used for storing files associated with an application. This bucket is associated with the application and can be used by the gcloud deployment commands. @OutputOnly

default_bucket

A Google Cloud Storage bucket which can be used by the application to store content. @OutputOnly

dispatch_rules

HTTP path dispatch rules for requests to the app that do not explicitly target a service or version. The rules are order-dependent. @OutputOnly

default_cookie_expiration

Determines the cookie expiration policy for the application. @OutputOnly

location

The location from which the application will be run. Application instances will run out of data centers in the chosen location and all of the application's End User Content will be stored at rest. The default is "us-central". Choices are: "us-central" - Central US "europe-west" - Western Europe "us-east1" - Eastern US

auth_domain

If set, only users from the specified Google Apps authentication domain may access the application. If not set, any Google Account may access the application.

id

The relative name/path of the application. Example: "myapp".