Struct google_appengine1_beta5::StaticFilesHandler

pub struct StaticFilesHandler {
    pub mime_type: Option<String>,
    pub expiration: Option<String>,
    pub path: Option<String>,
    pub upload_path_regex: Option<String>,
    pub require_matching_file: Option<bool>,
    pub http_headers: Option<HashMap<String, String>>,
    pub application_readable: Option<bool>,
}

Files served directly to the user for a given URL, such as images, CSS stylesheets, or JavaScript source files. Static file handlers describe which files in the application directory are static files, and which URLs serve them.

This type is not used in any activity, and only used as part of another schema.

Fields

mime_type

If specified, all files served by this handler will be served using the specified MIME type. If not specified, the MIME type for a file will be derived from the file's filename extension.

expiration

The length of time a static file served by this handler ought to be cached by web proxies and browsers.

path

The path to the static files matched by the URL pattern, from the application root directory. The path can refer to text matched in groupings in the URL pattern.

upload_path_regex

A regular expression that matches the file paths for all files that will be referenced by this handler.

require_matching_file

If true, this UrlMap entry does not match the request unless the file referenced by the handler also exists. If no such file exists, processing will continue with the next UrlMap that matches the requested URL.

http_headers

HTTP headers to use for all responses from these URLs.

application_readable

By default, files declared in static file handlers are uploaded as static data and are only served to end users, they cannot be read by an application. If this field is set to true, the files are also uploaded as code data so your application can read them. Both uploads are charged against your code and static data storage resource quotas.