Struct google_appengine1_beta5::Version

pub struct Version {
    pub basic_scaling: Option<BasicScaling>,
    pub beta_settings: Option<HashMap<String, String>>,
    pub manual_scaling: Option<ManualScaling>,
    pub creation_time: Option<String>,
    pub vm: Option<bool>,
    pub inbound_services: Option<Vec<String>>,
    pub instance_class: Option<String>,
    pub deployment: Option<Deployment>,
    pub error_handlers: Option<Vec<ErrorHandler>>,
    pub serving_status: Option<String>,
    pub id: Option<String>,
    pub threadsafe: Option<bool>,
    pub api_config: Option<ApiConfigHandler>,
    pub nobuild_files_regex: Option<String>,
    pub network: Option<Network>,
    pub handlers: Option<Vec<UrlMap>>,
    pub health_check: Option<HealthCheck>,
    pub default_expiration: Option<String>,
    pub name: Option<String>,
    pub libraries: Option<Vec<Library>>,
    pub deployer: Option<String>,
    pub env: Option<String>,
    pub disk_usage_bytes: Option<String>,
    pub automatic_scaling: Option<AutomaticScaling>,
    pub runtime: Option<String>,
    pub env_variables: Option<HashMap<String, String>>,
    pub resources: Option<Resources>,
}

A Version is a specific set of source code and configuration files deployed to a service.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

services versions create apps (request)

services versions get apps (response)

services versions patch apps (request)

Fields

basic_scaling

A service with basic scaling will create an instance when the application receives a request. The instance will be turned down when the app becomes idle. Basic scaling is ideal for work that is intermittent or driven by user activity.

beta_settings

Beta settings supplied to the application via metadata.

manual_scaling

A service with manual scaling runs continuously, allowing you to perform complex initialization and rely on the state of its memory over time.

creation_time

Creation time of this version. This will be between the start and end times of the operation that creates this version. @OutputOnly

vm

Whether to deploy this app in a VM container.

inbound_services

Before an application can receive email or XMPP messages, the application must be configured to enable the service.

instance_class

The instance class to use to run this app. Valid values for AutomaticScaling are [F1, F2, F4, F4_1G]. Valid values for ManualScaling and BasicScaling are [B1, B2, B4, B8, B4_1G]. Default: "F1" for AutomaticScaling, "B1" for ManualScaling and BasicScaling

deployment

Code and application artifacts that make up this version. Only returned in GET requests if view=FULL is set. May only be set on create requests; once created, is immutable.

error_handlers

Custom static error pages instead of these generic error pages, (limit 10 KB/page) Only returned in GET requests if view=FULL is set. May only be set on create requests; once created, is immutable.

serving_status

The current serving status of this version. Only SERVING versions will have instances created or billed for. If this field is unset when a version is created, SERVING status will be assumed. It is an error to explicitly set this field to SERVING_STATUS_UNSPECIFIED.

id

The relative name/path of the Version within the service. Example: "v1". Version specifiers can contain lowercase letters, digits, and hyphens. It cannot begin with the prefix ah- and the names default and latest are reserved and cannot be used.

threadsafe

If true, multiple requests can be dispatched to the app at once.

api_config

Serving configuration for Google Cloud Endpoints. Only returned in GET requests if view=FULL is set. May only be set on create requests; once created, is immutable.

nobuild_files_regex

Go only. Files that match this pattern will not be built into the app. May only be set on create requests.

network

Used to specify extra network settings (for VM runtimes only).

handlers

An ordered list of URL Matching patterns that should be applied to incoming requests. The first matching URL consumes the request, and subsequent handlers are not attempted. Only returned in GET requests if view=FULL is set. May only be set on create requests; once created, is immutable.

health_check

Configure health checking for the VM instances. Unhealthy VM instances will be stopped and replaced with new instances. Only returned in GET requests if view=FULL is set. May only be set on create requests; once created, is immutable.

default_expiration

The length of time a static file served by a static file handler ought to be cached by web proxies and browsers, if the handler does not specify its own expiration. Only returned in GET requests if view=FULL is set. May only be set on create requests; once created, is immutable.

name

The full path to the Version resource in the API. Example: "apps/myapp/services/default/versions/v1". @OutputOnly

libraries

Configuration for Python runtime third-party libraries required by the application. Only returned in GET requests if view=FULL is set. May only be set on create requests; once created, is immutable.

deployer

The email address of the user who created this version. @OutputOnly

env

The App Engine execution environment to use for this version. Default: "1"

disk_usage_bytes

Total size of version files hosted on App Engine disk in bytes. @OutputOnly

automatic_scaling

Automatic scaling is the scaling policy that App Engine has used since its inception. It is based on request rate, response latencies, and other application metrics.

runtime

The desired runtime. Values can include python27, java7, go, etc.

env_variables

Environment variables made available to the application. Only returned in GET requests if view=FULL is set. May only be set on create requests; once created, is immutable.

resources

Used to specify how many machine resources an app version needs (for VM runtimes only).