Struct google_appengine1_beta5::Application

pub struct Application {
    pub default_hostname: Option<String>,
    pub name: Option<String>,
    pub code_bucket: Option<String>,
    pub default_bucket: Option<String>,
    pub dispatch_rules: Option<Vec<UrlDispatchRule>>,
    pub default_cookie_expiration: Option<String>,
    pub location: Option<String>,
    pub auth_domain: Option<String>,
    pub id: Option<String>,
}

An Application resource contains the top-level configuration of an App Engine application.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

create apps (request)

get apps (response)

Fields
default_hostname: Option<String>

Hostname used to reach the application, as resolved by App Engine. @OutputOnly

name: Option<String>

Full path to the Application resource in the API. Example: apps/myapp. @OutputOnly

code_bucket: Option<String>

A Google Cloud Storage bucket that can be used for storing files associated with this application. This bucket is associated with the application and can be used by the gcloud deployment commands. @OutputOnly

default_bucket: Option<String>

A Google Cloud Storage bucket that can be used by the application to store content. @OutputOnly

dispatch_rules: Option<Vec<UrlDispatchRule>>

HTTP path dispatch rules for requests to the application that do not explicitly target a service or version. Rules are order-dependent. @OutputOnly

default_cookie_expiration: Option<String>

Cookie expiration policy for this application. @OutputOnly

location: Option<String>

Location from which this application will be run. Application instances will run out of data centers in the chosen location, which is also where all of the application's end user content is stored. Defaults to us-central. Options are: us-central - Central US europe-west - Western Europe us-east1 - Eastern US

auth_domain: Option<String>

Google Apps authentication domain that controls which users can access this application. Defaults to open access for any Google Account.

id: Option<String>

Identifier of the Application resource. This identifier is equivalent to the project ID of the Google Cloud Platform project where you want to deploy your application. Example: myapp.